# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restaurant Management System API",
        "description": "API documentation for Restaurant Management System Backend",
        "contact": {
            "email": "admin@restaurant.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Restaurant Management API Server"
        }
    ],
    "paths": {
...
        "/api/auth/dishes": {
            "get": {
                "tags": [
                    "Dishes"
                ],
                "summary": "Lấy danh sách món ăn",
                "operationId": "8b827576cf8d763c76d07686236a187e",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Trang hiện tại",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Số item mỗi trang",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10,
                            "maximum": 100
                        }
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Tìm theo tên món (partial match)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Bibimbap"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách món ăn"
                    }
                }
            },
            "post": {
                "tags": [
                    "Dishes"
                ],
                "summary": "Tạo mới món ăn",
                "operationId": "40ff5d1ea32db934545b843765e5dee1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "price",
                                    "category_id"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float"
                                    },
                                    "desc": {
                                        "type": "string"
                                    },
                                    "category_id": {
                                        "type": "string"
                                    },
                                    "cooking_time": {
                                        "type": "integer"
                                    },
                                    "image": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Dish created successfully"
                    }
                }
            }
        },
        "/api/auth/dishes/{id}": {
            "put": {
                "tags": [
                    "Dishes"
                ],
                "summary": "Cập nhật món ăn",
                "operationId": "2b089d40c47e816fb96673c164745283",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string"
                                    },
                                    "price": {
                                        "type": "number"
                                    },
                                    "desc": {
                                        "type": "string"
                                    },
                                    "category_id": {
                                        "type": "string"
                                    },
                                    "cooking_time": {
                                        "type": "integer"
                                    },
                                    "image": {
                                        "type": "string"
                                    },
                                    "is_active": {
                                        "type": "boolean"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Dish updated successfully"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Dishes"
                ],
                "summary": "Xóa món ăn",
                "operationId": "7b19bc5760d57ec8004bb78854315774",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Dish deleted successfully"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Enter JWT token in format: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}